libsocket: comparison of integers of different signs
../inet/sourcefilter.c:88:11: error: comparison of integers of different signs: 'int' and 'uint_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
        cpsize = MIN(orig_numsrc, *numsrc) * sizeof (struct sockaddr_storage);
                 ^   ~~~~~~~~~~~  ~~~~~~~
/code/illumos-gate/proto/root_i386/usr/include/sys/sysmacros.h:57:24: note: expanded from macro 'MIN'
                          ~  ^  ~
../inet/sourcefilter.c:184:11: error: comparison of integers of different signs: 'int' and 'uint32_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
        cpsize = MIN(orig_numsrc, *numsrc) * sizeof (struct in_addr);
                 ^   ~~~~~~~~~~~  ~~~~~~~
/code/illumos-gate/proto/root_i386/usr/include/sys/sysmacros.h:57:24: note: expanded from macro 'MIN'
                          ~  ^  ~
2 errors generated.

../socket/_soutil.c:217:18: error: comparison of integers of different signs: 'unsigned int' and 'int' [-Werror,-Wsign-compare]
                            semantics == type &&
                            ~~~~~~~~~ ^  ~~~~
../socket/_soutil.c:231:18: error: comparison of integers of different signs: 'unsigned int' and 'int' [-Werror,-Wsign-compare]
                            semantics == type) {
                            ~~~~~~~~~ ^  ~~~~
2 errors generated.

../inet/ruserpass.c:163:17: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
        if (t < 0 || t >= sizeof (buf)) {
                     ~ ^  ~~~~~~~~~~~~
1 error generated.

../inet/inet6_opt.c:162:25: error: comparison of integers of different signs: 'int' and 'socklen_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
        if ((offset + padbytes > extlen) || !extbuf) {
             ~~~~~~~~~~~~~~~~~ ^ ~~~~~~
1 error generated.

../inet/getnameinfo.c:230:16: error: comparison of integers of different signs: 'socklen_t' (aka 'unsigned int') and 'int' [-Werror,-Wsign-compare]
                        if (servlen < portlen + 1)
                            ~~~~~~~ ^ ~~~~~~~~~~~
../inet/getnameinfo.c:254:17: error: comparison of integers of different signs: 'socklen_t' (aka 'unsigned int') and 'int' [-Werror,-Wsign-compare]
                                if (servlen < portlen + 1)
                                    ~~~~~~~ ^ ~~~~~~~~~~~
../inet/getnameinfo.c:341:10: error: comparison of integers of different signs: 'int' and 'size_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
                        if (n >= zonelen) {
                            ~ ^  ~~~~~~~
3 errors generated.

../inet/getaddrinfo.c:584:34: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
                if (((zonestr - hostname) + 1) > sizeof (_hostname)) {
                     ~~~~~~~~~~~~~~~~~~~~~~~~  ^ ~~~~~~~~~~~~~~~~~~
1 error generated.

+++ usr/src/lib/libsocket/inet/getaddrinfo.c
@@ -581,7 +581,7 @@ get_addr(int family, const char *hostname, struct addrinfo *aip, struct
 	 */
 	if ((zonestr = strchr(hostname, '%')) != NULL) {
 		/* make sure we have room for <addr> portion of hostname */
-		if (((zonestr - hostname) + 1) > sizeof (_hostname)) {
+		if (((size_t)(zonestr - hostname) + 1) > sizeof (_hostname)) {
 			return (EAI_MEMORY);
 		}
